Consolidated all common functions for testing sorting algorithms into a single file
