+++ algorithms/Sorting/InsertionSort/test/InsertionSortTest.cpp
@@ -1,4 +1,5 @@
 #include "../InsertionSort.h"
+#include "../../../util/SortingUtil.h"
 #include "gtest/gtest.h"
 
 #define ARRAY_SIZE 1000
@@ -9,62 +10,6 @@ int main(int argc, char** argv)
 
     testing::InitGoogleTest(&argc, argv);
     return RUN_ALL_TESTS();
-}
-
-bool valueIsInArray(int arr[], const int& value, const size_t& arrSize)
-{
-    for (int i = 0; i < arrSize; i++) {
-        if (arr[i] == value)
-            return true;
-    }
-
-    return false;
-}
-
-bool valueIsInPointerArray(int* arr[], const int& value, const size_t& arrSize)
-{
-    for (int i = 0; i < arrSize; i++) {
-        if (*arr[i] == value)
-            return true;
-    }
-
-    return false;
-}
-
-bool arrayIsSorted(int arr[], const size_t& arrSize)
-{
-    int prev = arr[0];
-    for (int i = 1; i < arrSize; i++) {
-        if (arr[i] < prev)
-            return false;
-        prev = arr[i];
-    }
-
-    return true;
-}
-
-bool pointerArrayIsSorted(int* arr[], const size_t& arrSize)
-{
-    int prev = *arr[0];
-    for (int i = 1; i < arrSize; i++) {
-        if (*arr[i] < prev)
-            return false;
-        prev = *arr[i];
-    }
-
-    return true;
-}
-
-bool vectorIsSorted(vector<int> arr)
-{
-    int prev = arr[0];
-    for (int i = 1; i < arr.size(); i++) {
-        if (arr[i] < prev)
-            return false;
-        prev = arr[i];
-    }
-
-    return true;
 }
 
 TEST(InsertSort, ArraySort)
